refactor: RESTful api
