+++ src/main/java/cn/edu/nju/madpill/controller/UserController.java
@@ -2,20 +2,16 @@ package cn.edu.nju.madpill.controller;
 
 import cn.binarywang.wx.miniapp.api.WxMaService;
 import cn.binarywang.wx.miniapp.bean.WxMaJscode2SessionResult;
-import cn.edu.nju.madpill.domain.User;
 import cn.edu.nju.madpill.dto.Result;
-import cn.edu.nju.madpill.dto.TagDTO;
-import cn.edu.nju.madpill.exception.ExceptionSuppliers;
 import cn.edu.nju.madpill.service.UserService;
 import me.chanjar.weixin.common.error.WxErrorException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
-
-import static cn.edu.nju.madpill.utils.MadPillConstant.HEADER_MADPILL_TOKEN_KEY;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @author Aneureka
@@ -57,32 +53,4 @@ public class UserController {
                     .build();
         }
     }
-
-    @GetMapping(path = "/groups")
-    public Result getGroup(@RequestHeader(name = HEADER_MADPILL_TOKEN_KEY) String token) {
-        Optional<User> curUser = userService.getUserByToken(token);
-        if (curUser.isPresent()) {
-            return Result.builder()
-                    .data(userService.getGroups(curUser.get()))
-                    .code(HttpStatus.OK.value())
-                    .build();
-        } else {
-            throw ExceptionSuppliers.INVALID_TOKEN.get();
-        }
-    }
-
-    @PostMapping(path = "/groups")
-    public Result createGroup(@RequestHeader(name = HEADER_MADPILL_TOKEN_KEY) String token, @RequestBody String name) {
-        System.out.println(token  + "   " + name);
-        Optional<User> curUser = userService.getUserByToken(token);
-        if (curUser.isPresent()) {
-            return Result.builder()
-                    .data(userService.newGroup(name, curUser.get(), false))
-                    .code(HttpStatus.OK.value())
-                    .build();
-        } else {
-            throw ExceptionSuppliers.INVALID_TOKEN.get();
-        }
-    }
-
 }
