+++ src/main/java/cn/edu/nju/madpill/service/UserService.java
@@ -1,34 +1,20 @@
 package cn.edu.nju.madpill.service;
 
 import cn.edu.nju.madpill.config.CodecConfig;
-import cn.edu.nju.madpill.custommapper.UserAssistantMapper;
-import cn.edu.nju.madpill.domain.Group;
-import cn.edu.nju.madpill.domain.Tag;
 import cn.edu.nju.madpill.domain.User;
-import cn.edu.nju.madpill.domain.UserGroup;
-import cn.edu.nju.madpill.dto.GroupBriefDTO;
-import cn.edu.nju.madpill.dto.TagDTO;
-import cn.edu.nju.madpill.mapper.UserGroupMapper;
 import cn.edu.nju.madpill.mapper.UserMapper;
 import cn.edu.nju.madpill.utils.Base64XORCodec;
-import org.apache.ibatis.annotations.InsertProvider;
-import org.mybatis.dynamic.sql.insert.render.InsertStatementProvider;
 import org.mybatis.dynamic.sql.render.RenderingStrategies;
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.Generated;
-import java.time.*;
-import java.util.Date;
-import java.util.List;
+import java.time.LocalDateTime;
 import java.util.Optional;
 
-import static cn.edu.nju.madpill.custommapper.GroupAssistantDynamicSqlSupport.buildInsert;
-import static cn.edu.nju.madpill.mapper.GroupDynamicSqlSupport.group;
 import static cn.edu.nju.madpill.mapper.UserDynamicSqlSupport.user;
-import static cn.edu.nju.madpill.mapper.UserGroupDynamicSqlSupport.userGroup;
-import static org.mybatis.dynamic.sql.SqlBuilder.*;
+import static org.mybatis.dynamic.sql.SqlBuilder.isEqualTo;
+import static org.mybatis.dynamic.sql.SqlBuilder.select;
 
 /**
  * @author Aneureka
@@ -40,16 +26,15 @@ import static org.mybatis.dynamic.sql.SqlBuilder.*;
 @Transactional
 public class UserService {
 
+    private final GroupService groupService;
+
     private final UserMapper userMapper;
-    private final UserGroupMapper userGroupMapper;
-    private final UserAssistantMapper userAssistantMapper;
 
     private final CodecConfig codecConfig;
 
-    public UserService(UserMapper userMapper, UserAssistantMapper userAssistantMapper, UserGroupMapper userGroupMapper, CodecConfig codecConfig) {
+    public UserService(GroupService groupService, UserMapper userMapper, CodecConfig codecConfig) {
+        this.groupService = groupService;
         this.userMapper = userMapper;
-        this.userAssistantMapper = userAssistantMapper;
-        this.userGroupMapper = userGroupMapper;
         this.codecConfig = codecConfig;
     }
 
@@ -60,7 +45,7 @@ public class UserService {
             record.setOpenId(openId);
             record.setCreatedAt(LocalDateTime.now());
             userMapper.insert(record);
-            newGroup("我的药箱", record, false);
+            groupService.newGroup("我的药箱", record, false);
         }
     }
 
@@ -94,34 +79,5 @@ public class UserService {
         } else {
             throw new IllegalArgumentException("Invalid token.");
         }
-    }
-
-    public List<GroupBriefDTO> getGroups(User curUser) {
-        Long selectedId = curUser.getId();
-        SelectStatementProvider selectStatementProvider = select(group.id.as("group_id"), group.name.as("group_name"))
-                .from(userGroup)
-                .join(group)
-                .on(userGroup.groupId, equalTo(group.id))
-                .where(userGroup.userId, isEqualTo(selectedId))
-                .build()
-                .render(RenderingStrategies.MYBATIS3);
-
-        return userAssistantMapper.selectGroups(selectStatementProvider);
-    }
-
-    public Long newGroup(String name, User curUser, boolean firstGroup) {
-        Group group = new Group();
-        group.setCreateAt(LocalDateTime.now(Clock.system(ZoneId.of("Asia/Shanghai"))));
-        group.setCreateBy(curUser.getId());
-        group.setName(name);
-        group.setCanDelete(firstGroup);
-
-        userAssistantMapper.insert(buildInsert(group));
-        UserGroup userGroup = new UserGroup();
-        userGroup.setGroupId(group.getId());
-        userGroup.setUserId(group.getCreateBy());
-        userGroup.setAlias(name);
-        userGroupMapper.insert(userGroup);
-        return group.getId();
     }
 }
